new tests commit

+++ src/main/java/com/ontop/assessment/payment/infrastructure/adapters/outbound/wallet/TransactionRequest.java
@@ -5,10 +5,12 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class TransactionRequest {
     private final String userId;
-    private final long amount;
+    private final BigDecimal amount;
 }
